pyproject.toml
[tool.poetry]
name = "abstract-topic-identification"
version = "0.1.0"
description = "NSF abstracts topic identification"
authors = ["Roger Galindo <rogherschrodinger@gmail.com>"]
readme = "README.md"
packages = [{include = "abstract_topic_identification"}]

[tool.poetry.dependencies]
python = "^3.11"
nltk = "^3.8.1"
scikit-learn = "^1.5.0"
hydra-core = "^1.3.2"
dvc = "^3.51.0"
transformers = "^4.41.1"
pyldavis = "^3.4.1"
matplotlib = "^3.9.0"
umap-learn = "^0.5.6"
hdbscan = "^0.8.36"
streamlit = "^1.35.0"



[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
